Revert "fix: Ensure file uploads use RFC 7578 standard for multipart arrays"
The changes on the Speakeasy side don't seem to be working as intended.
Specifically, `multipartArrayFormat: standard` does not cause
`src/glean/api_client/utils/forms.py` to be updated upon the next
re-generation.

This brings back the manual work around for now, while we debug what is
going on with the generation issue. The way the original fix works is
that if we end up with fields that end with `[]` they are corrected,
which means our work around is safe to have alongside the fixed
generation (because once the generation is fixed there won't be any
fields with `[]`).

This reverts commit cbf4a87535f6d921370d0bb42b7ce5f9dbd930fa.